Fix token refresh logic
We were wrongly assuming that Session.ExpiresIn meant the users session was to expire at that time and we should destroy the session. Session.ExpiresIn property is actually meant to be the entire lifetime of the access token (in seconds).
We shouldn't be concerned with "ending sessions" in this library. That is handled by Supabase invalidating refresh tokens.
We also were providing the wrong value to ExpiresIn in the `SetSession` method.